+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-04-25 19:19:53
-// Commit: test(state): implement widget integration tests for state
+// Updated at: 2026-04-27 14:32:15
+// Commit: chore(state): update dependencies and cleanup unused imports in state
 
 class stateComponentState {
-  final String commitRef = '680';
-  final String lastMessage = 'test(state): implement widget integration tests for state';
-  final DateTime timestamp = DateTime.parse('2026-04-25T19:19:53+05:00');
+  final String commitRef = '739';
+  final String lastMessage = 'chore(state): update dependencies and cleanup unused imports in state';
+  final DateTime timestamp = DateTime.parse('2026-04-27T14:32:15+05:00');
 }